refactor(arp): replace magic numbers with defined constants for clarity

+++ libraries/pico-net-stack/inc/arp.h
@@ -15,9 +15,9 @@ struct arp
 };
 
 int8_t arp_is_request_for_me(
-    const struct arp* self, uint8_t* frame, uint16_t size, bool* state);
+    const struct arp* self, uint8_t* frame, uint8_t size, bool* state);
 
-int8_t arp_reply_request(const struct arp* self, uint8_t* frame, uint16_t size);
+int8_t arp_reply_request(const struct arp* self, uint8_t* frame, uint8_t size);
 
 /* ========================================================================== */
